ucg_service/pyproject.toml
[tool.poetry]
name = "ugc-sprint-1"
version = "0.1.0"
description = ""
authors = ["Schislyaev <80824032+Schislyaev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
clickhouse-driver = "^0.2.5"
pydantic = {extras = ["dotenv"], version = "^1.10.4"}
backoff = "^2.2.1"
fastapi = "^0.91.0"
pytest = "^7.2.1"
uvicorn = "^0.20.0"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.1"
kafka-python = "^2.0.2"
aiohttp = "^3.8.3"
aiokafka = "^0.8.0"
grpcio-tools = "^1.51.1"
sentry-sdk = {extras = ["fastapi"], version = "^1.15.0"}
flake8-html = "^0.4.3"
motor = "^3.1.1"
pyjwt = "^2.6.0"
orjson = "^3.8.6"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest-asyncio = "^0.20.3"
pytest = "^7.2.1"
mypy = "^1.0.1"


[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.20.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
